Updated folder structure, tests, json files, paver script, Removed old json files and test code
