+++ .gitignore
@@ -1,2 +1,3 @@
 */local.log
 */**.pyc
+**/.DS_Store
