--- android/config/local.json
@@ -1,16 +1,0 @@
-{
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "behave-browserstack",
-    "name": "local_test",
-    "browserstack.debug": true,
-    "browserstack.local": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "Google Pixel"
-  }]
-}
