--- android/config/parallel.json
@@ -1,24 +1,0 @@
-{
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "behave-browserstack",
-    "name": "parallel_test",
-    "browserstack.debug": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "Samsung Galaxy S8"
-  },
-  {
-    "device": "Samsung Galaxy S7"
-  },
-  {
-    "device": "Google Nexus 5"
-  },
-  {
-    "device": "Google Pixel"
-  }]
-}
