+++ android/examples/run-first-test/config.json
@@ -1,0 +1,13 @@
+{
+    "username":"BROWSERSTACK_USERNAME",
+    "access_key":"BROWSERSTACK_ACCESS_KEY",
+
+    "capabilities": {
+        "project": "First Behave Android Project",
+        "build": "Behave Android",
+        "name": "first_test",
+        "browserstack.debug": true,
+        "app": "bs://d318ec55142bf92b1fec9fce6904109294db6678",
+        "device": "Google Pixel"
+    }
+}
