+++ android/examples/run-parallel-tests/features/parallel_test.feature
@@ -1,0 +1,4 @@
+Feature: Wikipedia Search Functionality
+    Scenario: can find search results
+        Given I open the app and search for keyword "BrowserStack"
+        Then Search results should appear
