--- ios/config/parallel.json
@@ -1,17 +1,0 @@
-{
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "behave-browserstack",
-    "name": "parallel_test",
-    "browserstack.debug": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "iPhone 7"
-  }, {
-    "device": "iPhone 7 Plus"
-  }]
-}
