--- ios/config/single.json
@@ -1,15 +1,0 @@
-{
-  "user": "BROWSERSTACK_USERNAME",
-  "key": "BROWSERSTACK_ACCESS_KEY",
-
-  "capabilities": {
-    "build": "behave-browserstack",
-    "name": "single_test",
-    "browserstack.debug": true,
-    "app": "bs://<hashed app-id>"
-  },
-
-  "environments": [{
-    "device": "iPhone 7"
-  }]
-}
