+++ ios/examples/run-parallel-tests/features/parallel_test.feature
@@ -1,0 +1,5 @@
+Feature: 
+    Scenario: Displayed Text should match Input Text
+        Given I open the app and click on Text Button
+        Then Type "hello@browserstack.com" and hit enter
+        Then Verify displayed text matches input text
